Add a kind of console interface where one can type commands
This is done using a popular 'linenoise' library which is statically
compiled (and forked). I sitll do not have internet access, so I've
taken the code from the phantomjs code which I do have locally.

For now the read-eval-print-loop needs to run in another thread, as the
main thread is resposible for painting the QWebEnginePage.