Implement message receiving
This is just the first iteration of message receiving and it will miss
loads of messages.

Every chat monitors the message list and keeps an internal copy of all
the messages in its chat. Whenever the message list is updated, each
chat checks if it should update its message list, does a diff and sends
the appropriate signals.

The currentChat currently only changes when sending someone a message.
So only when you send that person a message will you get messages.