Reload the messsageList when we have unread items in a chat
This works by selecting that chat and reloading its message list.
however, during that time someone might be trying to send a message and
since the current chat would have changed, the message would be sent to
the wrong person, and that is just EVIL.

We need some kind of serializations that blocks other operations when
one operation is going on.